# Speed up recursions