--- a/src/nemos/basis/_basis.py
+++ b/src/nemos/basis/_basis.py
@@ -4,8 +4,10 @@ from __future__ import annotations
 import abc
 import copy
 from collections import OrderedDict
+from copy import deepcopy
 from functools import wraps
 from typing import Callable, Generator, Literal, Optional, Tuple, Union
+from contextlib import nullcontext
 
 import jax
 import numpy as np
@@ -25,16 +27,25 @@ def _bisect_mul(base: Basis, mul: int):
     if mul == 1:
         return base  # Base case
     half = _bisect_mul(base, mul // 2)
-    sum = half + half  # Less deep copying
-    return sum + base if mul % 2 else sum
+    context = getattr(half, "_set_shallow_copy_temporarily", nullcontext)
+    with context(True):
+        sum = half + half  # Less deep copying
+    with sum._set_shallow_copy_temporarily(True):
+        sum =  sum + base if mul % 2 else sum
+    return sum
 
 
 def _bisect_power(base: Basis, exp: int):
     if exp == 1:
         return base  # Base case
     half = _bisect_power(base, exp // 2)
-    squared = half * half  # Less deep copying
-    return squared * base if exp % 2 else squared
+    context = getattr(half, "_set_shallow_copy_temporarily", nullcontext)
+
+    with context(True):
+        squared = half * half  # Less deep copying
+    with squared._set_shallow_copy_temporarily(True):
+        squared = squared * base if exp % 2 else squared
+    return squared
 
 
 def add_docstring(method_name, cls):
@@ -511,7 +522,7 @@ class Basis(Base, abc.ABC, BasisTransformerMixin):
                     "Basis multiplication error. Integer multiplicative factor must be positive, "
                     f"{other} provided instead."
                 )
-            return _bisect_mul(self, other)
+            return _bisect_mul(self.__sklearn_clone__(), other)
         if not isinstance(other, Basis):
             raise TypeError(
                 "Basis multiplicative factor should be a Basis object or a positive integer!"
@@ -547,7 +558,7 @@ class Basis(Base, abc.ABC, BasisTransformerMixin):
         if exponent <= 0:
             raise ValueError("Basis exponent should be a non-negative integer!")
 
-        return _bisect_power(self, exponent)
+        return _bisect_power(self.__sklearn_clone__(), exponent)
 
     def __repr__(self):
         return format_repr(self)