--- a/src/nemos/basis/_basis.py
+++ b/src/nemos/basis/_basis.py
@@ -703,6 +703,10 @@ class Basis(Base, abc.ABC, BasisTransformerMixin):
                     "Basis multiplication error. Integer multiplicative factor must be positive, "
                     f"{other} provided instead."
                 )
+            elif not all(b._has_default_label for _, b in generate_basis_label_pair(self)):
+                raise ValueError(
+                    "Cannot multiply by an integer a basis including a user-defined labels."
+                )
             return _bisect_mul(self, other)
         if not isinstance(other, Basis):
             raise TypeError(