--- a/src/nemos/basis/_basis.py
+++ b/src/nemos/basis/_basis.py
@@ -19,6 +19,7 @@ from ..typing import FeatureMatrix
 from ..utils import format_repr, row_wise_kron
 from ..validation import check_fraction_valid_samples
 from ._basis_mixin import BasisTransformerMixin, CompositeBasisMixin
+from ._composition_utils import _recompute_all_default_labels
 
 
 def add_docstring(method_name, cls):
@@ -494,16 +495,24 @@ class Basis(Base, abc.ABC, BasisTransformerMixin):
                 b._has_default_label for _, b in generate_basis_label_pair(self)
             ):
                 raise ValueError(
-                    "Cannot multiply by an integer a basis including a user-defined labels (because then they won't be unique). Set labels after multiplication."
+                    "Cannot multiply by an integer a basis including a user-defined labels "
+                    "(because then they won't be unique). Set labels after multiplication."
                 )
             # default case
             if other == 1:
-                return deepcopy(self)
+                bas = deepcopy(self)
+                # reset parent
+                # takes care of case: bas.basis1 * 1
+                bas._parent = None
+                _recompute_all_default_labels(bas)
+                return bas
 
+            # parent is set to None at init for add and updated for self.
             add = AdditiveBasis(self, self)
-            with add._set_shallow_copy_temporarily(True):
+            with add._set_shallow_copy(True):
                 for _ in range(2, other):
                     add += deepcopy(self)
+            _recompute_all_default_labels(add)
             return add
 
         if not isinstance(other, Basis):
@@ -543,12 +552,18 @@ class Basis(Base, abc.ABC, BasisTransformerMixin):
 
         # default case
         if exponent == 1:
-            return deepcopy(self)
+            bas = deepcopy(self)
+            # reset parent
+            # takes care of case: bas.basis1 * 1
+            bas._parent = None
+            _recompute_all_default_labels(bas)
+            return bas
 
         mul = MultiplicativeBasis(self, self)
-        with mul._set_shallow_copy_temporarily(True):
+        with mul._set_shallow_copy(True):
             for _ in range(2, exponent):
                 mul *= deepcopy(self)
+        _recompute_all_default_labels(mul)
         return mul
 
     def __repr__(self):