… src/nemos/basis/_basis_mixin.py
… tests/test_basis.py
… tests/test_transformer_basis.py