In this PR:
- New accept integer for `__mul__`: `basis * 3` will be equivalent to `basis + basis+basis`
- Add an `__rmul__` enabling right multiplication`n * basis`
- Improved the efficiency of `__pow__` by bisect multiply (log2(n) multiplication, less deep copy)
- Add a mechanism that do not perform a deep-copy when cloning a bases

**Old clone performance**
```
bas = nmo.basis.BSplineEval(5)**500
%timeit bas.__sklearn_clone__()
2.47 s ± 107 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

**New clone performance**
```
bas = nmo.basis.BSplineEval(5)**500
%timeit bas.__sklearn_clone__()
10.9 ms ± 22 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
**Old Power Performance**
```
bas = nmo.basis.BSplineEval(5)
%timeit bas ** 500
2.49 s ± 76.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

**New Power Performance**
```
bas = nmo.basis.BSplineEval(5)
%timeit bas ** 500
8.46 ms ± 79.3 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```