@billbrod 2025-03-05
While reviewing this PR, I looked into how to use not-quite-global variables, like the new shallow_copy flag. That flag affects the behavior of some functions of **all** composite classes, so it can't be instance level (i.e., `self._shallow_copy`). Right now, it's class level (`self.__class__._shallow_copy`), which I think is right. While looking into this, came across [this thread](https://discuss.python.org/t/context-manager-that-stores-a-temporary-context-for-globals/43365/4), which is discussing how stdlib's `warnings` handles filters, and they're set at the module-level, using `sys.modules["warnings"].filters`. So if we had flags that affected the behavior across the module (not just within a single class/mixin), then that would be the way to go.